Storybook: Extract Vite 8 Babel fallback
