+++ storybook/main.ts
@@ -1,11 +1,6 @@
 import path from 'node:path';
 import { fileURLToPath } from 'node:url';
 import {
-	type ParserOptions,
-	transformAsync as transformWithBabel,
-} from '@babel/core';
-import {
-	createFilter,
 	type InlineConfig,
 	type PluginOption,
 	mergeConfig,
@@ -15,6 +10,7 @@ import react from '@vitejs/plugin-react';
 import type { StorybookConfig } from '@storybook/react-vite';
 import dsTokenFallbacks from '@wordpress/theme/postcss-plugins/postcss-ds-token-fallbacks';
 import dsTokenFallbacksJs from '@wordpress/theme/vite-plugins/vite-ds-token-fallbacks';
+import babel from './vite-babel-plugin.js';
 
 /**
  * @see https://storybook.js.org/docs/faq#how-do-i-fix-module-resolution-in-special-environments
@@ -26,7 +22,6 @@ function getAbsolutePath( packageName: string ) {
 }
 
 const { NODE_ENV = 'development' } = process.env;
-const shouldTransformWithBabel = createFilter( /\.[jt]sx?$/, /node_modules/ );
 
 const stories = [
 	'./stories/playground/**/*.story.@(jsx|tsx)',
@@ -134,58 +129,21 @@ const config: StorybookConfig = {
 				dsTokenFallbacksJs(),
 				react() as PluginOption,
 				// @rolldown/plugin-babel requires Node 22, but Gutenberg still
-				// supports Node 20. Run the existing Emotion transform directly.
-				{
-					name: 'transform-emotion-with-babel',
-					async transform( code: string, id: string ) {
-						const [ filePath ] = id.split( '?' );
-						if (
-							! shouldTransformWithBabel( filePath ) ||
-							! code.includes( '@emotion/' )
-						) {
-							return null;
-						}
-
-						const parserPlugins: NonNullable<
-							ParserOptions[ 'plugins' ]
-						> = [];
-						if ( ! filePath.endsWith( '.ts' ) ) {
-							parserPlugins.push( 'jsx' );
-						}
-						if ( /\.tsx?$/.test( filePath ) ) {
-							parserPlugins.push( 'typescript' );
-						}
-
-						const result = await transformWithBabel( code, {
-							babelrc: false,
-							configFile: false,
-							filename: id,
-							generatorOpts: {
-								decoratorsBeforeExport: true,
-								importAttributesKeyword: 'with',
-							},
-							parserOpts: {
-								allowAwaitOutsideFunction: true,
-								plugins: parserPlugins,
-								sourceType: 'module',
-							},
-							plugins: [
-								getAbsolutePath( '@emotion/babel-plugin' ),
-							],
-							retainLines:
-								NODE_ENV !== 'production' &&
-								filePath.endsWith( 'x' ),
-							sourceFileName: filePath,
-							sourceMaps: true,
-						} );
-
-						if ( ! result?.code ) {
-							return null;
-						}
-
-						return { code: result.code, map: result.map };
+				// supports Node 20. Keep the same call shape so this fallback can
+				// be replaced with the package after the Node upgrade.
+				await babel( {
+					generatorOpts: {
+						decoratorsBeforeExport: true,
+						importAttributesKeyword: 'with',
 					},
-				},
+					overrides: [
+						{
+							test: /x(?:$|\?)/,
+							retainLines: NODE_ENV !== 'production',
+						},
+					],
+					plugins: [ getAbsolutePath( '@emotion/babel-plugin' ) ],
+				} ),
 				{
 					name: 'load-js-files-as-jsx',
 					enforce: 'pre',
